+++ Commands/sparkle_help/sparkle_instances_help.py
@@ -205,6 +205,18 @@ def _get_list_instance(instances_source: str):
 	return list_instance
 
 
+def get_instance_list_from_path(path: Path) -> list[str]:
+	"""Return a list of instance name strings located in a given path."""
+	# Multi-file instances
+	if _check_existence_of_instance_list_file(str(path)):
+		list_all_filename = _get_list_instance(str(path))
+	# Single file instances
+	else:
+		list_all_filename = sfh.get_list_all_cnf_filename(str(path))
+
+	return list_all_filename
+
+
 def _copy_instance_list_to_reference(instances_source: Path):
 	instance_list_path = Path(instances_source / Path(__sparkle_instance_list_file))
 	target_path = Path(sgh.reference_list_dir / Path(instances_source.name + sgh.instance_list_postfix))