+++ Commands/test/all.sh
@@ -92,6 +92,9 @@ Commands/test/validate_configured_vs_default.sh
 # Test ablation run configured vs default
 Commands/test/run_ablation.sh
 
+# Test run configured solver
+Commands/test/run_configured_solver.sh
+
 # Test construct parallel portfolio
 Commands/test/construct_sparkle_parallel_portfolio.sh