+++ Examples/Resources/Solvers/PbO-CCSAT-Generic/sparkle_smac_wrapper.py
@@ -5,6 +5,7 @@ import os
 import time
 import random
 import sys
+from pathlib import Path
 
 
 def get_time_pid_random_string():
@@ -79,4 +80,9 @@ for line in output_list:
 
 print(r'Result for SMAC: ' + status + r', ' + str(run_time) + r', 0, 0, ' + str(seed))
 
+if specifics == 'rawres':
+	raw_result_path = Path(runsolver_watch_data_path.replace('.log', '.rawres'))
 
+	with raw_result_path.open('w') as outfile:
+		for line in output_list:
+			outfile.write(line)